requirements.txt
google-api-python-client==2.118.0
PyGithub==2.3.0
python-dateutil==2.9.0
PyYAML==6.0.2
youtube-transcript-api>=1.0.0
groq>=0.9.0
yt-dlp>=2024.1.0
